Changed 'int id' => 'long id' in calculate function signature (t4v4.0 upgrade).

+++ trunk/bond_risk_simple/objects/ir_curve.c
@@ -13,7 +13,7 @@
 
 using namespace std;
 
-int calculate_ir_curve( obj_loc_t obj_loc, int id )
+int calculate_ir_curve( obj_loc_t obj_loc, long id )
 {
     int current_period_start(0);
     int current_period_end(0);
